projects/contextual-ai/api/requirements.txt
uvicorn
fastapi
dotenv
google-cloud
google-cloud-storage
google-cloud-bigquery
google-genai
google-cloud-aiplatform
python-dotenv
